2008 removal of iss parts.

+++ source2/Debug/Cosmos.Debug.Common/CDebugger/DebugSymbol.cs
@@ -34,31 +34,52 @@ namespace Cosmos.Debug.Common.CDebugger
 
 	public class MLDebugSymbol {
 
-        protected static void CreateCPDB(string aPathname) {
-            File.Delete(aPathname);
+        static protected FbConnection DBConn;
 
+        protected static void OpenCPDB(string aPathname, bool aCreate) {
             var xCSB = new FbConnectionStringBuilder();
             xCSB.ServerType = FbServerType.Embedded;
             xCSB.Database = aPathname;
+            xCSB.UserID = "sysdba";
+            xCSB.Password = "masterkey";
+            xCSB.Pooling = false;
 
-            FbConnection.CreateDatabase(xCSB.ToString());
-
-            using (var xConn = new FbConnection(xCSB.ToString())) {
-                var xExec = new FbBatchExecution(xConn);
-                //foreach (string cmd in script.Results) {
-                //    xExec.SqlStatements.Add(cmd);
-                //}
-                //fbe.Execute();
+            if (aCreate) {
+                File.Delete(aPathname);
+                FbConnection.CreateDatabase(xCSB.ToString());
             }
+
+            DBConn = new FbConnection(xCSB.ToString());
+            DBConn.Open();
+        }
+
+        protected static void CreateCPDB(string aPathname) {
+            OpenCPDB(aPathname, true);
+            var xExec = new FbBatchExecution(DBConn);
+
+            xExec.SqlStatements.Add(
+                "CREATE TABLE SYMBOL ("
+                + "   LABELNAME   VARCHAR(255)  NOT NULL"
+                + " , ADDRESS     BIGINT        NOT NULL"
+                + " , STACKDIFF   INT           NOT NULL"
+                + " , ILASMFILE   VARCHAR(255)  NOT NULL"
+                + " , TYPETOKEN   INT           NOT NULL"
+                + " , METHODTOKEN INT           NOT NULL"
+                + " , ILOFFSET    INT           NOT NULL"
+                + " , METHODNAME  VARCHAR(255)  NOT NULL"
+                + ");"
+            );
+
+            xExec.Execute();
+            // Batch execution closes the connection, so we have to reopen it
+            DBConn.Open();
         }
 
 		public static void WriteSymbolsListToFile(IEnumerable<MLDebugSymbol> aSymbols, string aFile) {
             CreateCPDB(Path.ChangeExtension(aFile, ".cpdb"));
 
             var xDS = new SymbolsDS();
-            // This is a dataset and XML currently.
-            // Later change it to a real DB as we don't need it in RAM all at once but we need it to be fast.
-            // Probably embedded Firebird
+            // Is a real DB now, but we still store all in RAM. We dont need to. Need to change to query DB as needed instead.
             foreach(var xItem in aSymbols){
                 var x = xDS.Entry.NewEntryRow();
                 x.LabelName = xItem.LabelName;
@@ -70,11 +91,25 @@ namespace Cosmos.Debug.Common.CDebugger
                 x.ILOffset = xItem.ILOffset;
                 x.MethodName = xItem.MethodName;
                 xDS.Entry.AddEntryRow(x);
+
+                var xCmd = new FbCommand("INSERT INTO SYMBOL (LABELNAME, ADDRESS, STACKDIFF, ILASMFILE, TYPETOKEN, METHODTOKEN, ILOFFSET, METHODNAME)"
+                    + " VALUES (@LABELNAME, @ADDRESS, @STACKDIFF, @ILASMFILE, @TYPETOKEN, @METHODTOKEN, @ILOFFSET, @METHODNAME)"
+                    , DBConn);
+                xCmd.Parameters.Add("@LABELNAME", xItem.LabelName);
+                xCmd.Parameters.Add("@ADDRESS", xItem.Address);
+                xCmd.Parameters.Add("@STACKDIFF", xItem.StackDifference);
+                xCmd.Parameters.Add("@ILASMFILE", xItem.AssemblyFile);
+                xCmd.Parameters.Add("@TYPETOKEN", xItem.TypeToken);
+                xCmd.Parameters.Add("@METHODTOKEN", xItem.MethodToken);
+                xCmd.Parameters.Add("@ILOFFSET", xItem.ILOffset);
+                xCmd.Parameters.Add("@METHODNAME", xItem.MethodName);
+                xCmd.ExecuteNonQuery();
             }
             xDS.WriteXml(aFile);
 		}
 
 		public static void ReadSymbolsListFromFile(List<MLDebugSymbol> aSymbols, string aFile) {
+            OpenCPDB(Path.ChangeExtension(aFile, ".cpdb"), false);
             var xDS = new SymbolsDS();
             xDS.ReadXml(aFile);
             foreach (SymbolsDS.EntryRow x in xDS.Entry.Rows) {
